+++ db/migrate/20160919220536_create_purchases.rb
@@ -2,7 +2,7 @@ class CreatePurchases < ActiveRecord::Migration
   def change
     create_table :purchases do |t|
       t.string :item
-      t.integer :cost
+      t.decimal :cost
       t.text :note
       t.integer :plan_id
